Refactored auth, now falls back to direct SMTP
Email login is now "always on" even if not configured.

This is not as reliable as using a dedicated service as outbound port 25 may be blocked from some servers and because mail from servers that have not been configured as mail hosts (e.g. with corresponding SPF records in DNS) may be more likely to be blocked or flagged as spam.

Support for additional third party mail service providers (Gmail, mailgun, etc) to follow.