Email address verification prompt added
This initial implementation is not configurable and may be slightly annoying.

A user is still not required to verify their email address before being able to post but this lays the ground work for that option, which is a way of helping to reduce spam.

It's also lays some of the groundwork for  'closed forums'.